+++ lib/view_component/slot.rb
@@ -3,5 +3,11 @@
 module ViewComponent
   class Slot
     attr_accessor :content
+
+    def with_content(this_content)
+      self.content = this_content
+
+      self
+    end
   end
 end